Add support for Google Optimize MVP
This is a first step towards #1. Not too happy with the implementation,
but it works, and will probably need more changes to the overall
architecture when implementing support for other platforms anyway. This
is good enough for now, as it allows us to start focusing on the UX, as
well as testing some edge cases.