Preparing for user-based SRM based on GA data
Added variables to capture all the necessary info for generating the GA link. What's left is to use a proper custom report URL instead of the one I used as suggested by @heinrich333 , and then write code to extract the data from the report and perform a SRM on it. Perhaps this is not something we can do using another iframe.