Added VWO support
Everything seemed to work well when testing the most crucial things. 

Note to self: Find a way to get the data we need jumping through less hoops, as this is prone to death by DOM structure changes. If I had to use a class from their HTML to describe the queries I used, it would be ng-dirty.

We also should consider:
• Moving the iframe function outside of platform sections, might need it for more than two platforms soon
• Experiment with conditionals for running the interval, so we don't spam the reports of people using JS error tracking needlessly

[Getting indentation bullied in 3... 2... 1... Just kidding, please feel free to fix as you see fit.]