Fixed JS error when iframeforweight was not available
Across all platforms. I cancel the SRM check if the weights iframe is not available - presumption is this means the extension is not looking at a test results page at that point in time. A recheck will happen in 1 sec and if the URL has changed, all will proceed as usual.