Add a basic status indicator badge
First step towards providing users with more feedback about the SRM
check performed (e.g. #5 and/or #6).

I find it strange that I needed to add browser_action for this, but
there seems to be no way around this. I don't think this actually gives
the extension more permissions, but documentation is a bit unclear.